Added bounds checking
WIP - need to write a type resolver for expectations (segment and event) so that the coreect subtype is deserialied.

Also adds appoximation forms to interval.

Adds docs for intervals.

Renamed props on centroid so that they are consistent with bounds and time range.

Started merging of expect and expect_events into one list.

Added the concept of an error expectation/ Used when data can not be found as opposed to data not matching expected values.