Changed the way false positives are counted
Also changed expectation result so that the producer of the result can change what type of result is produced. This is useful in cases where a segment-level analysis needs to produce event-level failures.

Which leads us to, the no_events expectation no produces one negative *event* result ( a false positive) for every event produced in a segment, as well as the traditional *segment* level failure.